+++ src/mlinsights/helpers/__init__.py
@@ -1,0 +1,6 @@
+"""
+@file
+@brief Shortcuts to *helpers*.
+"""
+
+from .visualize import pipeline2dot